+++ ProgramFunctions/Fields.fnc.php
@@ -383,13 +383,11 @@ function GetFieldsForm( $table, $title, $RET, $extra_category_fields = [], $type
 
 	if ( $id )
 	{
-		// FJ field name required.
 		$header .= '<td>' . MLTextInput(
 			issetVal( $RET['TITLE'], '' ),
 			'tables[' . $id . '][TITLE]',
-			( empty( $RET['TITLE'] ) ? '<span class="legend-red">' : '' ) . _( 'Field Name' ) .
-				( empty( $RET['TITLE'] ) ? '</span>' : '' ),
-			'maxlength="200"'
+			_( 'Field Name' ),
+			'required maxlength="200"'
 		) . '</td>';
 
 		if ( ! $type_options )
@@ -560,9 +558,8 @@ function GetFieldsForm( $table, $title, $RET, $extra_category_fields = [], $type
 		$header .= '<td>' . MLTextInput(
 			issetVal( $RET['TITLE'], '' ),
 			'tables[' . $category_id . '][TITLE]',
-			( empty( $RET['TITLE'] ) ? '<span class="legend-red">' : '') . _( 'Title' ) .
-				( empty( $RET['TITLE'] ) ? '</span>' : '' ),
-			'maxlength="36"'
+			_( 'Title' ),
+			'required maxlength="36"'
 		) . '</td>';
 
 		// Sort Order field.